+++ Changelog.md
@@ -4,6 +4,7 @@
 
 - Added isolated tests for shifting mechanics in USA calendars - previously untested (#335).
 - Added Barbados by @ludsoft.
+- Added Berlin specific holidays (#340).
 
 ## v4.2.0 (2019-02-21)